refactor(lib::playlist): xspf tests to xspf module
